app/pyproject.toml
[project]
name = "sample-app"
version = "0.1.0"
description = "sample app"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.61.0",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "ty>=0.0.14",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]
